requirement.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
libretranslatepy==2.1.1
lxml==5.2.1
polib==1.2.0
python-dotenv==1.0.1
requests==2.31.0
translate==3.6.1
urllib3==2.2.1
